Completes intro and homepage merge final phase.
Merged the content and homepage completely, Added particles.js to
the homepage for better design. Also shifted the js and static files
in respecting folders, thus giving the project a cleaner structure.

The next task would be to finalize the individual pages for each of the
options in the content view and working on it's overall design and
improving it.